refactor: move and rename TermsAndConditionsCheckbox
