--- src/components/DownloadAndShareComponents/TermsAndConditionsCheckbox/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./TermsAndConditionsCheckbox";
